+++ doc/release-notes.md
@@ -9,5 +9,8 @@ Network upgrade
 
 At the MTP time of `1700049600` (November 15, 2023 12:00:00 UTC), the following
 changes will become activated:
- - The miner fund ratio is increased to 32% of the coinbase reward
-
+ - The miner fund ratio is increased to 32% of the coinbase reward.
+ - The staking reward output is added to the coinbase transaction. It sends 10%
+   of the block reward to a payout address selected from the avalanche network.
+   Blocks that don't include or have an incorrect output will be rejected by the
+   avalanche network.